APIADD: (in last checkin) VistaTransformMatrix::Set/GetAdjunct(Translation|BaseMatrix) overloads for double
git-svn-id: https://svn.rwth-aachen.de/repos/vrgroup-svn/projects/Vista/trunk@39441 33c38b33-09d2-4561-8d4d-1bfa2f58f885

+++ VistaCoreLibs/VistaBase/VistaTransformMatrix.h
@@ -773,9 +773,9 @@ inline void VistaTransformMatrix::SetTranslation( const float a3fTranslation[3] 
 }
 inline void VistaTransformMatrix::SetTranslation( const double a3dTranslation[3] )
 {
-	operator()( 0, 3 ) = (double)a3dTranslation[0];
-	operator()( 1, 3 ) = (double)a3dTranslation[1];
-	operator()( 2, 3 ) = (double)a3dTranslation[2];
+	operator()( 0, 3 ) = (float)a3dTranslation[0];
+	operator()( 1, 3 ) = (float)a3dTranslation[1];
+	operator()( 2, 3 ) = (float)a3dTranslation[2];
 }
 inline void VistaTransformMatrix::SetTranslation( const VistaVector3D& v3Translation )
 {
